Extract UserInfo, WebAppMiddlewareConfiguration, and WebAppMiddlewareExtensions in Shared Kernel to separate files
